Feature/modify username setup UI css (#94)
* Modify CSS/UI for UsernameSetup to align with SplashPage

* Modify CSS for UsernameSetup to match Splash

* Modify UI/css of UsernameSetup for intuitive username registration, uncluttered UI

* Username input text box outlined in red when error

* feat(username-validation): Add client-side username length validation with visual feedback

Implement client-side validation for username input in the UsernameSetup component to ensure the length is between 3 and 100 characters. Upon validation failure, the input field now visually indicates an error state with a red border. This enhancement improves user feedback and input experience.

* feat(username-validation): Add client-side username length validation with visual feedback

Implement client-side validation for username input in the UsernameSetup component to ensure the length is between 3 and 100 characters. Upon validation failure, the input field now visually indicates an error state with a red border. This enhancement improves user feedback and input experience.